@@ -4,6 +4,7 @@
 
 // ignore_for_file: public_member_api_docs
 
+import 'dart:developer' as developer;
 import 'dart:isolate';
 import 'dart:math';
 import 'dart:ui';
@@ -25,7 +26,6 @@ final ReceivePort port = ReceivePort();
 SharedPreferences prefs;
 
 Future<void> main() async {
-  // TODO(bkonyi): uncomment
   WidgetsFlutterBinding.ensureInitialized();
 
   // Register the UI isolate's SendPort to allow for communication from the
@@ -38,15 +38,17 @@ Future<void> main() async {
   if (!prefs.containsKey(countKey)) {
     await prefs.setInt(countKey, 0);
   }
-  runApp(AlarmManagerExampleApp());
+  runApp(const AlarmManagerExampleApp());
 }
 
 /// Example app for Espresso plugin.
 class AlarmManagerExampleApp extends StatelessWidget {
+  const AlarmManagerExampleApp({Key key}) : super(key: key);
+
   // This widget is the root of your application.
   @override
   Widget build(BuildContext context) {
-    return MaterialApp(
+    return const MaterialApp(
       title: 'Flutter Demo',
       home: _AlarmHomePage(title: 'Flutter Demo Home Page'),
     );
@@ -54,7 +56,7 @@ class AlarmManagerExampleApp extends StatelessWidget {
 }
 
 class _AlarmHomePage extends StatefulWidget {
-  _AlarmHomePage({Key key, this.title}) : super(key: key);
+  const _AlarmHomePage({Key key, this.title}) : super(key: key);
   final String title;
 
   @override
@@ -75,8 +77,7 @@ class _AlarmHomePageState extends State<_AlarmHomePage> {
   }
 
   Future<void> _incrementCounter() async {
-    print('Increment counter!');
-
+    developer.log('Increment counter!');
     // Ensure we've loaded the updated count from the background isolate.
     await prefs.reload();
 
@@ -90,8 +91,7 @@ class _AlarmHomePageState extends State<_AlarmHomePage> {
 
   // The callback for our alarm
   static Future<void> callback() async {
-    print('Alarm fired!');
-
+    developer.log('Alarm fired!');
     // Get the previous cached count and increment it.
     final prefs = await SharedPreferences.getInstance();
     final currentCount = prefs.getInt(countKey);
@@ -126,13 +126,13 @@ class _AlarmHomePageState extends State<_AlarmHomePage> {
                 ),
                 Text(
                   prefs.getInt(countKey).toString(),
-                  key: ValueKey('BackgroundCountText'),
+                  key: const ValueKey('BackgroundCountText'),
                   style: textStyle,
                 ),
               ],
             ),
             ElevatedButton(
-              key: ValueKey('RegisterOneShotAlarm'),
+              key: const ValueKey('RegisterOneShotAlarm'),
               onPressed: () async {
                 await AndroidAlarmManager.oneShot(
                   const Duration(seconds: 5),
@@ -143,7 +143,7 @@ class _AlarmHomePageState extends State<_AlarmHomePage> {
                   wakeup: true,
                 );
               },
-              child: Text(
+              child: const Text(
                 'Schedule OneShot Alarm',
               ),
             ),
