@@ -26,6 +26,11 @@ typedef UPowerDeviceFactory = UPowerDevice Function();
 
 /// The Linux implementation of BatteryPlatform.
 class BatteryPlusLinux extends BatteryPlatform {
+  /// Register this dart class as the platform implementation for linux
+  static void registerWith() {
+    BatteryPlatform.instance = BatteryPlusLinux();
+  }
+
   /// Returns the current battery level in percent.
   @override
   Future<int> get batteryLevel {
@@ -50,7 +55,7 @@ class BatteryPlusLinux extends BatteryPlatform {
   StreamController<BatteryState>? _stateController;
 
   @visibleForTesting
-  // ignore: public_member_api_docs
+  // ignore: public_member_api_docs, prefer_function_declarations_over_variables
   UPowerDeviceFactory createDevice = () => UPowerDevice.display();
 
   void _addState(UPowerBatteryState value) {
