@@ -3,12 +3,8 @@
 // found in the LICENSE file.
 
 import 'dart:async';
-import 'dart:io' show Platform;
 
-import 'package:flutter/foundation.dart' show kIsWeb;
 import 'package:connectivity_plus_platform_interface/connectivity_plus_platform_interface.dart';
-import 'package:connectivity_plus_platform_interface/method_channel_connectivity.dart';
-import 'package:connectivity_plus_linux/connectivity_plus_linux.dart';
 
 // Export enums from the platform_interface so plugin users can use them directly.
 export 'package:connectivity_plus_platform_interface/connectivity_plus_platform_interface.dart'
@@ -31,27 +27,15 @@ class Connectivity {
 
   static Connectivity? _singleton;
 
-  static bool _manualDartRegistrationNeeded = true;
-
-  // This is to manually endorse Dart implementations until automatic
-  // registration of Dart plugins is implemented. For details see
-  // https://github.com/flutter/flutter/issues/52267.
   static ConnectivityPlatform get _platform {
-    if (_manualDartRegistrationNeeded) {
-      // Only do the initial registration if it hasn't already been overridden
-      // with a non-default instance.
-      if (!kIsWeb &&
-          ConnectivityPlatform.instance is MethodChannelConnectivity) {
-        if (Platform.isLinux) {
-          ConnectivityPlatform.instance = ConnectivityLinux();
-        }
-      }
-      _manualDartRegistrationNeeded = false;
-    }
     return ConnectivityPlatform.instance;
   }
 
   /// Fires whenever the connectivity state changes.
+  ///
+  /// On iOS, the connectivity status might not update when WiFi
+  /// status changes, this is a known issue that only affects simulators.
+  /// For details see https://github.com/fluttercommunity/plus_plugins/issues/479.
   Stream<ConnectivityResult> get onConnectivityChanged {
     return _platform.onConnectivityChanged;
   }
