@@ -1,29 +1,18 @@
-// Copyright 2019 Google LLC
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//      http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
+// Copyright 2019 The Chromium Authors. All rights reserved.
+// Use of this source is governed by a BSD-style license that can
+// be found in the LICENSE file.
 
 import Cocoa
 import FlutterMacOS
-import Network
-import SystemConfiguration.CaptiveNetwork
 
 public class ConnectivityPlugin: NSObject, FlutterPlugin, FlutterStreamHandler {
-  var eventSink: FlutterEventSink?
-  var pathMonitor = NWPathMonitor()
+  private let connectivityProvider: ConnectivityProvider
+  private var eventSink: FlutterEventSink?
 
-  override init() {
+  init(connectivityProvider: ConnectivityProvider) {
+    self.connectivityProvider = connectivityProvider
     super.init()
-    pathMonitor.pathUpdateHandler = pathUpdateHandler
+    self.connectivityProvider.connectivityUpdateHandler = connectivityUpdateHandler
   }
 
   public static func register(with registrar: FlutterPluginRegistrar) {
@@ -35,7 +24,14 @@ public class ConnectivityPlugin: NSObject, FlutterPlugin, FlutterStreamHandler {
       name: "dev.fluttercommunity.plus/connectivity_status",
       binaryMessenger: registrar.messenger)
 
-    let instance = ConnectivityPlugin()
+    let connectivityProvider: ConnectivityProvider
+    if #available(macOS 10.14, *) {
+      connectivityProvider = PathMonitorConnectivityProvider()
+    } else {
+      connectivityProvider = ReachabilityConnectivityProvider()
+    }
+
+    let instance = ConnectivityPlugin(connectivityProvider: connectivityProvider)
     streamChannel.setStreamHandler(instance)
 
     registrar.addMethodCallDelegate(instance, channel: channel)
@@ -44,28 +40,23 @@ public class ConnectivityPlugin: NSObject, FlutterPlugin, FlutterStreamHandler {
   public func handle(_ call: FlutterMethodCall, result: @escaping FlutterResult) {
     switch call.method {
     case "check":
-      result(statusFromPath(path: pathMonitor.currentPath))
+      result(statusFrom(connectivityType: connectivityProvider.currentConnectivityType))
     default:
       result(FlutterMethodNotImplemented)
     }
   }
 
-  /// Returns a string describing connection type
-  ///
-  /// - Parameters:
-  ///   - path: an instance of NWPath
-  /// - Returns: connection type string
-  private func statusFromPath(path: NWPath) -> String {
-    if path.status == .satisfied {
-        if path.usesInterfaceType(.wifi) {
-            return "wifi"
-        } else if path.usesInterfaceType(.cellular) {
-            return "mobile"
-        } else if path.usesInterfaceType(.wiredEthernet) {
-            return "ethernet"
-        }
+  private func statusFrom(connectivityType: ConnectivityType) -> String {
+    switch connectivityType {
+    case .wifi:
+      return "wifi"
+    case .cellular:
+      return "mobile"
+    case .wiredEthernet:
+      return "ethernet"
+    case .none:
+      return "none"
     }
-    return "none"
   }
 
   public func onListen(
@@ -73,16 +64,17 @@ public class ConnectivityPlugin: NSObject, FlutterPlugin, FlutterStreamHandler {
     eventSink events: @escaping FlutterEventSink
   ) -> FlutterError? {
     eventSink = events
-    pathMonitor.start(queue: .global(qos: .background))
+    connectivityProvider.start()
+    connectivityUpdateHandler(connectivityType: connectivityProvider.currentConnectivityType)
     return nil
   }
 
-  private func pathUpdateHandler(path: NWPath) {
-    eventSink?(statusFromPath(path: path))
+  private func connectivityUpdateHandler(connectivityType: ConnectivityType) {
+    eventSink?(statusFrom(connectivityType: connectivityType))
   }
 
   public func onCancel(withArguments _: Any?) -> FlutterError? {
-    pathMonitor.cancel()
+    connectivityProvider.stop()
     eventSink = nil
     return nil
   }
