@@ -5,6 +5,7 @@
 // ignore_for_file: public_member_api_docs
 
 import 'dart:async';
+import 'dart:developer' as developer;
 import 'dart:io';
 
 import 'package:network_info_plus/network_info_plus.dart';
@@ -22,10 +23,12 @@ void _enablePlatformOverrideForDesktop() {
 
 void main() {
   _enablePlatformOverrideForDesktop();
-  runApp(MyApp());
+  runApp(const MyApp());
 }
 
 class MyApp extends StatelessWidget {
+  const MyApp({Key? key}) : super(key: key);
+
   // This widget is the root of your application.
   @override
   Widget build(BuildContext context) {
@@ -34,13 +37,13 @@ class MyApp extends StatelessWidget {
       theme: ThemeData(
         primarySwatch: Colors.blue,
       ),
-      home: MyHomePage(title: 'Flutter Demo Home Page'),
+      home: const MyHomePage(title: 'Flutter Demo Home Page'),
     );
   }
 }
 
 class MyHomePage extends StatefulWidget {
-  MyHomePage({Key? key, this.title}) : super(key: key);
+  const MyHomePage({Key? key, this.title}) : super(key: key);
 
   final String? title;
 
@@ -93,7 +96,7 @@ class _MyHomePageState extends State<MyHomePage> {
         wifiName = await _networkInfo.getWifiName();
       }
     } on PlatformException catch (e) {
-      print(e.toString());
+      developer.log('Failed to get Wifi Name', error: e);
       wifiName = 'Failed to get Wifi Name';
     }
 
@@ -113,49 +116,49 @@ class _MyHomePageState extends State<MyHomePage> {
         wifiBSSID = await _networkInfo.getWifiBSSID();
       }
     } on PlatformException catch (e) {
-      print(e.toString());
+      developer.log('Failed to get Wifi BSSID', error: e);
       wifiBSSID = 'Failed to get Wifi BSSID';
     }
 
     try {
       wifiIPv4 = await _networkInfo.getWifiIP();
     } on PlatformException catch (e) {
-      print(e.toString());
+      developer.log('Failed to get Wifi IPv4', error: e);
       wifiIPv4 = 'Failed to get Wifi IPv4';
     }
 
     try {
       wifiIPv6 = await _networkInfo.getWifiIPv6();
     } on PlatformException catch (e) {
-      print(e.toString());
+      developer.log('Failed to get Wifi IPv6', error: e);
       wifiIPv6 = 'Failed to get Wifi IPv6';
     }
 
     try {
       wifiSubmask = await _networkInfo.getWifiSubmask();
     } on PlatformException catch (e) {
-      print(e.toString());
+      developer.log('Failed to get Wifi submask address', error: e);
       wifiSubmask = 'Failed to get Wifi submask address';
     }
 
     try {
       wifiBroadcast = await _networkInfo.getWifiBroadcast();
     } on PlatformException catch (e) {
-      print(e.toString());
+      developer.log('Failed to get Wifi broadcast', error: e);
       wifiBroadcast = 'Failed to get Wifi broadcast';
     }
 
     try {
       wifiGatewayIP = await _networkInfo.getWifiGatewayIP();
     } on PlatformException catch (e) {
-      print(e.toString());
+      developer.log('Failed to get Wifi gateway address', error: e);
       wifiGatewayIP = 'Failed to get Wifi gateway address';
     }
 
     try {
       wifiSubmask = await _networkInfo.getWifiSubmask();
     } on PlatformException catch (e) {
-      print(e.toString());
+      developer.log('Failed to get Wifi submask', error: e);
       wifiSubmask = 'Failed to get Wifi submask';
     }
 
